Develop

+++ app/Http/Controllers/BillController.php
@@ -178,7 +178,7 @@ class BillController extends Controller
      * @bodyParam last_name2 required No obligatorio. Segundo apellido de proveedor nuevo
      * @bodyParam country required No obligatorio. País de proveedor nuevo
      * @bodyParam state required No obligatorio. Provincia de proveedor nuevo
-     * @bodyParam city required No obligatorio. Cantón de proveedor nuevo
+     * @bodyParam city required No obligatorio. Cantón de provee dor nuevo
      * @bodyParam district required No obligatorio. Distrito de proveedor nuevo
      * @bodyParam neighborhood required No obligatorio. Barrio de proveedor nuevo
      * @bodyParam zip required No obligatorio. Código postal de proveedor nuevo
@@ -190,33 +190,35 @@ class BillController extends Controller
     public function store(Request $request)
     {
       
+        $company = currentCompanyModel();
         $request->validate([
             'subtotal' => 'required',
             'items' => 'required',
         ]);
-      
-        $bill = new Bill();
-        $company = currentCompanyModel();
-        $bill->company_id = $company->id;
-
-        //Datos generales y para Hacienda
-        $bill->document_type = "01";
-        $bill->hacienda_status = "03";
-        $bill->status = "02";
-        $bill->payment_status = "01";
-        $bill->payment_receipt = "";
-        $bill->generation_method = "M";
-        $bill->reference_number = $company->last_bill_ref_number + 1;
-        $bill->accept_status = 1;
-        
-        $bill->setBillData($request);
-        
-        $company->last_bill_ref_number = $bill->reference_number;
-        $company->save();
-        
-        clearBillCache($bill);
-      
-        return redirect('/facturas-recibidas');
+        if(CalculatedTax::validarMes($request->generated_date)){
+            $bill = new Bill();
+            $bill->company_id = $company->id;
+            //Datos generales y para Hacienda
+            $bill->document_type = "01";
+            $bill->hacienda_status = "03";
+            $bill->status = "02";
+            $bill->payment_status = "01";
+            $bill->payment_receipt = "";
+            $bill->generation_method = "M";
+            $bill->reference_number = $company->last_bill_ref_number + 1;
+            $bill->accept_status = 1;
+            
+            $bill->setBillData($request);
+            
+            $company->last_bill_ref_number = $bill->reference_number;
+            $company->save();
+            
+            clearBillCache($bill);
+            return redirect('/facturas-recibidas');
+            
+        }else{
+            return redirect('/facturas-recibidas')->withError('Mes seleccionado ya fue cerrado');
+        }
     }
 
     /**
@@ -251,16 +253,23 @@ class BillController extends Controller
         $company = currentCompanyModel();
         
         $bill = Bill::findOrFail($id);
-        $units = UnidadMedicion::all()->toArray();
-        $this->authorize('update', $bill);
-      
-        //Valida que la factura recibida sea generada manualmente. De ser generada por XML o con el sistema, no permite edición.
-        if( $bill->generation_method != 'M' && $bill->generation_method != 'XLSX' ){
-          return redirect('/facturas-recibidas');
-        } 
-        $arrayActividades = $company->getActivities();
-      
+        $FechaEmision = explode(" ", $bill->generated_date);
+        $FechaEmision = explode("-", $FechaEmision[0]);
+        $FechaEmision = $FechaEmision[2]."/".$FechaEmision[1]."/".$FechaEmision[0];
+        if(CalculatedTax::validarMes($FechaEmision)){
+            $units = UnidadMedicion::all()->toArray();
+            $this->authorize('update', $bill);
+          
+            //Valida que la factura recibida sea generada manualmente. De ser generada por XML o con el sistema, no permite edición.
+            if( $bill->generation_method != 'M' && $bill->generation_method != 'XLSX' ){
+              return redirect('/facturas-recibidas');
+            } 
+            $arrayActividades = $company->getActivities();
+
         return view('Bill/edit', compact('bill', 'units', 'arrayActividades') );
+        }else{
+            return redirect('/facturas-recibidas')->withError('Mes seleccionado ya fue cerrado');
+        }
     }
 
     /**
@@ -275,17 +284,21 @@ class BillController extends Controller
     {
         
         $bill = Bill::findOrFail($id);
-        $this->authorize('update', $bill);
-      
-        //Valida que la factura emitida sea generada manualmente. De ser generada por XML o con el sistema, no permite edición.
-        if( $bill->generation_method != 'M' && $bill->generation_method != 'XLSX' ){
-          return redirect('/facturas-emitidas');
+        if(CalculatedTax::validarMes($request->generated_date)){
+            $this->authorize('update', $bill);
+          
+            //Valida que la factura emitida sea generada manualmente. De ser generada por XML o con el sistema, no permite edición.
+            if( $bill->generation_method != 'M' && $bill->generation_method != 'XLSX' ){
+              return redirect('/facturas-emitidas');
+            }
+          
+            $bill->setBillData($request);
+            
+            clearBillCache($bill);
+        
+        }else{
+            return redirect('/facturas-recibidas')->withError('Mes seleccionado ya fue cerrado');
         }
-      
-        $bill->setBillData($request);
-        
-        clearBillCache($bill);
-        
         return redirect('/facturas-recibidas');
     }
 
@@ -486,21 +499,29 @@ class BillController extends Controller
                     $xml = simplexml_load_string( file_get_contents($file) );
                     $json = json_encode( $xml ); // convert the XML string to JSON
                     $arr = json_decode( $json, TRUE );
-                    
-                    $identificacionReceptor = array_key_exists('Receptor', $arr) ? $arr['Receptor']['Identificacion']['Numero'] : 0;
-                    $identificacionEmisor = $arr['Emisor']['Identificacion']['Numero'];
-                    $consecutivoComprobante = $arr['NumeroConsecutivo'];
-                    $clave = $arr['Clave'];
-                    
-                    //Compara la cedula de Receptor con la cedula de la compañia actual. Tiene que ser igual para poder subirla
-                    if( preg_replace("/[^0-9]+/", "", $company->id_number) == preg_replace("/[^0-9]+/", "", $identificacionReceptor ) ) {
-                        //Registra el XML. Si todo sale bien, lo guarda en S3
-                        $bill = Bill::saveBillXML( $arr, 'XML' );
-                        if( $bill ) {
-                            Bill::storeXML( $bill, $file );
+
+                    $FechaEmision = explode("T", $arr['FechaEmision']);
+                    $FechaEmision = explode("-", $FechaEmision[0]);
+                    $FechaEmision = $FechaEmision[2]."/".$FechaEmision[1]."/".$FechaEmision[0];
+
+                    if(CalculatedTax::validarMes($FechaEmision)){
+                        $identificacionReceptor = array_key_exists('Receptor', $arr) ? $arr['Receptor']['Identificacion']['Numero'] : 0;
+                        $identificacionEmisor = $arr['Emisor']['Identificacion']['Numero'];
+                        $consecutivoComprobante = $arr['NumeroConsecutivo'];
+                        $clave = $arr['Clave'];
+                        
+                        //Compara la cedula de Receptor con la cedula de la compañia actual. Tiene que ser igual para poder subirla
+                        if( preg_replace("/[^0-9]+/", "", $company->id_number) == preg_replace("/[^0-9]+/", "", $identificacionReceptor ) ) {
+                            //Registra el XML. Si todo sale bien, lo guarda en S3
+                            $bill = Bill::saveBillXML( $arr, 'XML' );
+                            if( $bill ) {
+                                Bill::storeXML( $bill, $file );
+                            }
+                        }else{
+                            return back()->withError( "El documento $consecutivoComprobante no le pertenece a su compañía actual" );
                         }
                     }else{
-                        return back()->withError( "El documento $consecutivoComprobante no le pertenece a su compañía actual." );
+                        return back()->withError('Mes seleccionado ya fue cerrado');
                     }
                 }
             }
@@ -544,36 +565,44 @@ class BillController extends Controller
 
     public function validar($id){
         $current_company = currentCompany();
-        $company = Company::select('commercial_activities')->where('id', $current_company)->first();
-        $activities_company = explode(", ", $company->commercial_activities);
-        $commercial_activities = Actividades::whereIn('codigo', $activities_company)->get();
         $bill = Bill::find($id);
-        $codigos_etax = CodigoIvaSoportado::get();
-        $categoria_productos = ProductCategory::whereNotNull('bill_iva_code')->get();
+            $company = Company::select('commercial_activities')->where('id', $current_company)->first();
+            $activities_company = explode(", ", $company->commercial_activities);
+            $commercial_activities = Actividades::whereIn('codigo', $activities_company)->get();
+            $codigos_etax = CodigoIvaSoportado::get();
+            $categoria_productos = ProductCategory::whereNotNull('bill_iva_code')->get();
 
-        return view('Bill/validar', compact('bill', 'commercial_activities', 'codigos_etax', 'categoria_productos'));
+            return view('Bill/validar', compact('bill', 'commercial_activities', 'codigos_etax', 'categoria_productos'));
+        
     }
 
     public function GuardarValidar(Request $request)
     {
         $bill = Bill::findOrFail($request->bill);
         
-        $bill->activity_company_verification = $request->actividad_comercial;
-        $bill->is_code_validated = true;
-        foreach( $request->items as $item ) {
-            BillItem::where('id', $item['id'])
-            ->update([
-              'iva_type' =>  $item['iva_type'],
-              'product_type' =>  $item['product_type'],
-              'porc_identificacion_plena' =>  $item['porc_identificacion_plena']
-            ]);
-        }
-        
-        $bill->save();
-        
-        clearBillCache($bill);
+        $FechaEmision = explode(" ", $bill->generated_date);
+        $FechaEmision = explode("-", $FechaEmision[0]);
+        $FechaEmision = $FechaEmision[2]."/".$FechaEmision[1]."/".$FechaEmision[0];
+        if(CalculatedTax::validarMes($FechaEmision)){
+            $bill->activity_company_verification = $request->actividad_comercial;
+            $bill->is_code_validated = true;
+            foreach( $request->items as $item ) {
+                BillItem::where('id', $item['id'])
+                ->update([
+                  'iva_type' =>  $item['iva_type'],
+                  'product_type' =>  $item['product_type'],
+                  'porc_identificacion_plena' =>  $item['porc_identificacion_plena']
+                ]);
+            }
+            
+            $bill->save();
+            
+            clearBillCache($bill);
 
-        return back()->withMessage( 'La factura '. $bill->document_number . ' ha sido validada');
+            return back()->withMessage( 'La factura '. $bill->document_number . ' ha sido validada');
+        }else{
+            return back()->withError('Mes seleccionado ya fue cerrado');
+        }
 
     }
     
@@ -598,23 +627,30 @@ class BillController extends Controller
     public function confirmarValidacion( Request $request, $id )
     {
         $bill = Bill::findOrFail($id);
-        $this->authorize('update', $bill);
-        
-        $tipoIva = $request->tipo_iva;
-        foreach( $bill->items as $item ) {
-            $item->iva_type = $request->tipo_iva;
-            $item->save();
+        $FechaEmision = explode(" ", $bill->generated_date);
+        $FechaEmision = explode("-", $FechaEmision[0]);
+        $FechaEmision = $FechaEmision[2]."/".$FechaEmision[1]."/".$FechaEmision[0];
+        if(CalculatedTax::validarMes($FechaEmision)){
+            $this->authorize('update', $bill);
+            
+            $tipoIva = $request->tipo_iva;
+            foreach( $bill->items as $item ) {
+                $item->iva_type = $request->tipo_iva;
+                $item->save();
+            }
+            
+            $bill->is_code_validated = true;
+            $bill->save();
+            
+            if( $bill->year == 2018 ) {
+                clearLastTaxesCache($bill->company->id, 2018);
+            }
+            clearBillCache($bill);
+            
+            return redirect('/facturas-recibidas/aceptaciones')->withMessage( 'La factura '. $bill->document_number . 'ha sido validada');
+        }else{
+            return redirect('/facturas-recibidas/aceptaciones')->withError('Mes seleccionado ya fue cerrado');
         }
-        
-        $bill->is_code_validated = true;
-        $bill->save();
-        
-        if( $bill->year == 2018 ) {
-            clearLastTaxesCache($bill->company->id, 2018);
-        }
-        clearBillCache($bill);
-        
-        return redirect('/facturas-recibidas/aceptaciones')->withMessage( 'La factura '. $bill->document_number . 'ha sido validada');
     }
     
     /**
@@ -662,18 +698,27 @@ class BillController extends Controller
     public function authorizeBill ( Request $request, $id )
     {
         $bill = Bill::findOrFail($id);
-        $this->authorize('update', $bill);
+        $FechaEmision = explode(" ", $bill->generated_date);
+        $FechaEmision = explode("-", $FechaEmision[0]);
+        $FechaEmision = $FechaEmision[2]."/".$FechaEmision[1]."/".$FechaEmision[0];
+        if(CalculatedTax::validarMes($FechaEmision)){
         
-        if ( $request->autorizar ) {
-            $bill->is_authorized = true;
-            $bill->save();
-            return redirect('/facturas-recibidas/autorizaciones')->withMessage( 'La factura '. $bill->document_number . 'ha sido autorizada. Recuerde validar el código');
-        }else {
-            $bill->is_authorized = false;
-            $bill->is_void = true;
-            BillItem::where('bill_id', $bill->id)->delete();
-            $bill->delete();
-            return redirect('/facturas-recibidas/autorizaciones')->withMessage( 'La factura '. $bill->document_number . 'ha sido rechazada');
+            $this->authorize('update', $bill);
+            
+            if ( $request->autorizar ) {
+                $bill->is_authorized = true;
+                $bill->save();
+                return redirect('/facturas-recibidas/autorizaciones')->withMessage( 'La factura '. $bill->document_number . 'ha sido autorizada. Recuerde validar el código');
+            }else {
+                $bill->is_authorized = false;
+                $bill->is_void = true;
+                BillItem::where('bill_id', $bill->id)->delete();
+                $bill->delete();
+                return redirect('/facturas-recibidas/autorizaciones')->withMessage( 'La factura '. $bill->document_number . 'ha sido rechazada');
+            }
+
+        }else{
+            return redirect('/facturas-recibidas/autorizaciones')->withError('Mes seleccionado ya fue cerrado');
         }
     }
     
@@ -793,12 +838,21 @@ class BillController extends Controller
     public function markAsNotAccepted ( $id )
     {
         $bill = Bill::findOrFail($id);
-        $this->authorize('update', $bill);
+        $FechaEmision = explode(" ", $bill->generated_date);
+        $FechaEmision = explode("-", $FechaEmision[0]);
+        $FechaEmision = $FechaEmision[2]."/".$FechaEmision[1]."/".$FechaEmision[0];
+        if(CalculatedTax::validarMes($FechaEmision)){
         
-        $bill->accept_status = 0;
-        $bill->is_code_validated = false;
-        $bill->save();
-        return redirect('/facturas-recibidas/aceptaciones')->withMessage( 'La factura '. $bill->document_number . ' ha sido incluida para aceptación');
+            $this->authorize('update', $bill);
+            
+            $bill->accept_status = 0;
+            $bill->is_code_validated = false;
+            $bill->save();
+            return redirect('/facturas-recibidas/aceptaciones')->withMessage( 'La factura '. $bill->document_number . ' ha sido incluida para aceptación');
+        }else{
+            return redirect('/facturas-recibidas/aceptaciones')->withError('Mes seleccionado ya fue cerrado');
+        }
+
     }
     
     
@@ -850,18 +904,26 @@ class BillController extends Controller
     public function correctAccepted ( Request $request, $id )
     {
         $bill = Bill::findOrFail($id);
-        $this->authorize('update', $bill);
-        
-        $bill->accept_iva_condition = $request->accept_iva_condition ? $request->accept_iva_condition : '02';
-        $bill->accept_iva_acreditable = $request->accept_iva_acreditable;
-        $bill->accept_iva_gasto = $request->accept_iva_gasto;
-        $bill->accept_status = 1;
-        $bill->hacienda_status = "01";
-        $bill->accept_id_number = currentCompany();
+        $FechaEmision = explode(" ", $bill->generated_date);
+        $FechaEmision = explode("-", $FechaEmision[0]);
+        $FechaEmision = $FechaEmision[2]."/".$FechaEmision[1]."/".$FechaEmision[0];
+        if(CalculatedTax::validarMes($FechaEmision)){
+            $this->authorize('update', $bill);
+            
+            $bill->accept_iva_condition = $request->accept_iva_condition ? $request->accept_iva_condition : '02';
+            $bill->accept_iva_acreditable = $request->accept_iva_acreditable;
+            $bill->accept_iva_gasto = $request->accept_iva_gasto;
+            $bill->accept_status = 1;
+            $bill->hacienda_status = "01";
+            $bill->accept_id_number = currentCompany();
 
-        $bill->save();
-        clearBillCache($bill);
-        return redirect('/facturas-recibidas/aceptaciones')->withMessage( 'La factura '. $bill->document_number . 'ha sido aceptada');
+            $bill->save();
+            clearBillCache($bill);
+            return redirect('/facturas-recibidas/aceptaciones')->withMessage( 'La factura '. $bill->document_number . 'ha sido aceptada');
+        }else{
+            return redirect('/facturas-recibidas/aceptaciones')->withError('Mes seleccionado ya fue cerrado');
+        }
+
     }
 
     /**
